% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ata.R
\name{ata}
\alias{ata}
\title{Age to Age factors}
\usage{
ata(triangle)
}
\arguments{
\item{triangle}{Cumulated triangle as a matrix}
}
\value{
The age to age factors
}
\description{
\code{ata} computes the age to age coefficients for a triangle
}
\examples{
 links <- ata(triangleExampleEngland)

}
