% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtaImpact.R
\name{AtaImpact}
\alias{AtaImpact}
\title{Age to Age factors impacts}
\usage{
AtaImpact(triangle, mode = "value")
}
\arguments{
\item{triangle}{Cumulated triangle as a matrix}

\item{mode}{Character. The way to display the impact: as a value ("value") or as a percent of the total reserve ("percent").}
}
\value{
A triangle equivalent to the age-to-age factors triangle, but with the impact on the reserves instead.
}
\description{
\code{AtaImpact} computes the impact on the total reserve from removing each age to age factor in the Chain Ladder estimation for coefficents.
}
\examples{
 impact <- AtaImpact(triangleExampleEngland)

}
