% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapBenktander.R
\name{BootstrapBenktander}
\alias{BootstrapBenktander}
\title{Bootstrap}
\usage{
BootstrapBenktander(triangle, ultimateClaims, nBoot = 1000, weight = NA)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix.}

\item{ultimateClaims}{Ultimate claims by accident year.}

\item{nBoot}{Number of samples. Default value equals to 1000.}

\item{weight}{Boolean matrix with 1 row and 1 column less than the triangle to tell if the link ratio is to be considered: 1 for yes, 0 for no}
}
\value{
A list containing the following objects:
\itemize{
  \item{triangleList: the list of bootsraped triangles }
  \item{ibnrByAccidentYear: a data frame containing ibnrs by accident year}
  \item{ibnr: a vector of the total ibnr values)}
  \item{predictionError: estimation of the prediction error on IBNR}}
}
\description{
\code{BootstrapBenktander} computes the bootstraped distribution for IBNR on a Benktander method.
}
\examples{
ultimateClaims <- c(3901463,5433719,5378826,5297906,4858200,5111171,5660771,6784799,5642266,4969825)
bbt <- BootstrapBenktander(triangleExampleEngland, ultimateClaims, 1000)

}
