% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapBornFerg.R
\name{BootstrapBornFerg}
\alias{BootstrapBornFerg}
\title{Bootstrap}
\usage{
BootstrapBornFerg(triangle, ultimateClaims, nBoot = 1000, weight = NA)
}
\arguments{
\item{triangle}{Cumulated triangle as a matrix. The matrix should be square}

\item{ultimateClaims}{Ultimate claims by accident year.}

\item{nBoot}{Number of samples. Default value equals to 1000.}

\item{weight}{Boolean matrix the same size of the triangle to tell if the value is to be considered: 1 for yes, 0 for no. First column is not considered}
}
\value{
A list containing the following objects:
\itemize{
  \item{triangleList: the list of bootsraped triangles }
  \item{ibnrByAccidentYear: a data frame containing ibnrs by accident year}
  \item{ibnr: a vector of the total ibnr values)}
  \item{predictionError: estimation of the prediction error on IBNR}}
}
\description{
\code{BootstrapBornFerg} computes the bootstraped distribution for IBNR on a Bornhuetter Fergusson method.
}
\examples{
ultimateClaims <- c(3901463,5433719,5378826,5297906,4858200,5111171,5660771,6784799,5642266,4969825)
bbf <- BootstrapBornFerg(triangleExampleEngland, ultimateClaims, 1000)

}
