% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapChainLadder.R
\name{BootstrapChainLadder}
\alias{BootstrapChainLadder}
\title{Bootstrap}
\usage{
BootstrapChainLadder(triangle, nBoot = 1000, weight = NA)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix.}

\item{nBoot}{Number of samples. Default value equals to 1000.}

\item{weight}{Boolean matrix with 1 row and 1 column less than the triangle to tell if the link ratio is to be considered: 1 for yes, 0 for no}
}
\value{
A list containing the following objects:
\itemize{
  \item{triangleList: the list of bootsraped triangles }
  \item{ibnrByAccidentYear: a data frame containing ibnrs by accident year}
  \item{ibnr: a vector of the total ibnr values)}
  \item{predictionError: estimation of the prediction error on IBNR}}
}
\description{
\code{BootstrapChainLadder} computes the bootstraped distribution for IBNR with an ODP mehtod.
}
\examples{
bcl <- BootstrapChainLadder(triangleExampleEngland, 1000)

}
