% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapMack.R
\name{BootstrapMack}
\alias{BootstrapMack}
\title{Bootstrap Mack}
\usage{
BootstrapMack(triangle, nBoot = 1000)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix}

\item{nBoot}{Number of samples. Default value to 1000}
}
\value{
A list containing the following objects:
\itemize{
  \item{triangleList: the list of bootsraped triangles }
  \item{ibnrByAccidentYear: a data frame containing ibnrs by accident year}
  \item{ibnr: a vector of the total ibnr values)}
  \item{predictionError: estimation of the prediction error on IBNR}}
}
\description{
\code{BootstrapMack} computes the bootstraped distribution for IBNR with a Mack Model
}
\examples{
bm <- BootstrapMack(triangleExampleEngland, 1000)

}
