% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChainLadder.R
\name{ChainLadder}
\alias{ChainLadder}
\title{Chain Ladder}
\usage{
ChainLadder(triangle, weight = NA)
}
\arguments{
\item{triangle}{Cumulated triangle as a matrix}

\item{weight}{Boolean matrix the same size of the triangle to tell if the value is to be considered: 1 for yes, 0 for no. First column is not considered}
}
\value{
A list containing the following objects:
\itemize{
  \item{triangle: the input triangle }
  \item{developedTriangle: the developed triangle}
  \item{lambdas: the passing coefficients}
  \item{ibnrByAccidentYear: ibnr reserve by accident year}
  \item{ibnr: total ibnr reserve}}
}
\description{
\code{ChainLadder} applies the Chain Ladder method to a cumulated claim triangle.
}
\details{
Missing values are handled. They just need to be replaced by a NA.
}
\examples{
outputCL <- ChainLadder(triangleExampleEngland)

}
