% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChainLadderHyp.R
\name{ChainLadderHyp}
\alias{ChainLadderHyp}
\title{Verify chain ladder hypothes}
\usage{
ChainLadderHyp(triangle)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix}
}
\value{
a plot to confirm the hypothesis
}
\description{
\code{ChainLadderHype} verifies that there is no inflation by ploting the residuals by calendar year.
}
\examples{
ChainLadderHyp(triangleExampleEngland)

}
