% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cook.R
\name{Cook}
\alias{Cook}
\title{Cook's distance}
\usage{
Cook(triangle)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix}
}
\value{
a plot to displaying the Cook's distances
}
\description{
\code{Cook} plots Cook's distance for each age to age factor relatively to the linear model they belong.
}
\examples{
Cook(triangleExampleEngland)

}
