% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumulate.R
\name{Cumulate}
\alias{Cumulate}
\title{Cumulates a triangle}
\usage{
Cumulate(triangle)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix}
}
\value{
The cumulated triangle as a matrix
}
\description{
\code{Cumulate} Cumulates a triangle
}
\examples{
 decTriangle <- Decumulate(triangleExampleEngland)
 cumTriangle <- Cumulate(triangleExampleEngland)

}
