% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decumulate.R
\name{Decumulate}
\alias{Decumulate}
\title{Decumulates a triangle}
\usage{
Decumulate(triangle)
}
\arguments{
\item{triangle}{Cumulated triangle as a matrix}
}
\value{
The decumulated triangle as a matrix
}
\description{
\code{Decumulate} decumulates a triangle
}
\examples{
 decTriangle <- Decumulate(triangleExampleEngland)

}
