% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndepMack93.R
\name{IndepMack93}
\alias{IndepMack93}
\title{Mack93 independence}
\usage{
IndepMack93(triangle, range = 0.95)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix}

\item{range}{Range used for hypothesis testing.}
}
\value{
A list containing the following objects:
\itemize{
  \item{hypTab: the table for hypothesis testing }
  \item{valueTest: the values for hypothesis testing as a data.frame}
  \item{validated: boolean. If the test is validated with the selected range}}
}
\description{
\code{IndepMack93} test the hypothesis of the independence of the development factors across
calendar years.
}
\details{
Implementation of the technique for error estimation developed by Mack in its 1993 article "MEASURING THE VARIABILITY
OF CHAIN LADDER RESERVE ESTIMATES"
}
\examples{
error <- IndepMack93(triangleExampleEngland)

}
