% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{latest}
\alias{latest}
\title{Latest Observations}
\usage{
latest(triangle)
}
\arguments{
\item{triangle}{Cumulated triangle as a matrix}
}
\value{
The latest observation
}
\description{
\code{latest} extracts the latests observations (payments) from a triangle.
}
\examples{
 l <- latest(triangleExampleEngland)

}
