% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MackFirstHyp.R
\name{MackFirstHyp}
\alias{MackFirstHyp}
\title{Verify first Mack hypothesis}
\usage{
MackFirstHyp(triangle)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix}
}
\value{
a plot to confirm the hypothesis
}
\description{
\code{MackFirstHyp} verify first Mack hypothesis about the linearity of claims.
}
\examples{
MackFirstHyp(triangleExampleEngland)

}
