% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MackSecondHyp.R
\name{MackSecondHyp}
\alias{MackSecondHyp}
\title{Verify second Mack hypothesis}
\usage{
MackSecondHyp(triangle, weight = NA)
}
\arguments{
\item{triangle}{Cumulated triangle as a matrix}

\item{weight}{Boolean matrix the same size of the triangle to tell if the value is to be considered: 1 for yes, 0 for no. First column is not considered}
}
\value{
a plot to confirm the hypothesis
}
\description{
\code{MackSecondHyp} verifies second Mack hypothesis about the independance of residuals.
}
\examples{
MackSecondHyp(triangleExampleEngland)

}
