% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settlement.R
\name{Settlement}
\alias{Settlement}
\title{Settlement rate}
\usage{
Settlement(triangle)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix}
}
\value{
a plot
}
\description{
\code{Settlement} plots the settlement rate for each development year.
}
\examples{
Settlement(triangleExampleEngland)

}
