% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle-data.R
\docType{data}
\name{triangleExampleEngland}
\alias{triangleExampleEngland}
\title{England dataset}
\format{
An object of class \code{"matrix"}.
}
\source{
https://pdfs.semanticscholar.org/9093/f7faff74391892a9c61e4702697e888d9adb.pdf
}
\usage{
data(triangleExampleEngland)
}
\description{
Data from a paper on stochastic reserving in insurance
}
\examples{
data(triangleExampleEngland)
triangleExampleEngland
}
\references{
Stochastic Claims Reserving in General Insurance (2002)  P.D. England and R.J. Verrall
(\href{https://www.cambridge.org/core/journals/british-actuarial-journal/article/stochastic-claims-reserving-in-general-insurance/60026990B6A88E8A6DDEECABD6506C65}{England})
}
\keyword{datasets}
