% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BornFerg.R
\name{BornFerg}
\alias{BornFerg}
\title{Bornhuetter Fergusson}
\usage{
BornFerg(triangle, ultimateClaims)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix.}

\item{ultimateClaims}{Ultimate claims by accident year.}
}
\value{
A list containing the following objects:
\itemize{
  \item{triangle: the input triangle }
  \item{developedTriangle: the developed triangle}
  \item{gammas: the Bornhuetter Fergusson coefficients}
  \item{ibnrByAccidentYear: ibnr reserve by accident year}
  \item{ibnr: total ibnr reserve}}
}
\description{
\code{BornFerg} applies the Bornhuetter Fergusson method to a cumulated claim triangle.
}
\details{
Missing values are handled. There just need to be replaced by a NA.

It is possible to have more than one complete line.
}
\examples{
ultimateClaims <- c(3901463,5433719,5378826,5297906,4858200,5111171,5660771,6784799,5642266,4969825)
outputBF <- BornFerg(triangleExampleEngland, ultimateClaims)

}
