% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChainLadder.R
\name{ChainLadder}
\alias{ChainLadder}
\title{Chain Ladder}
\usage{
ChainLadder(triangle, weight = NA)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix}

\item{weight}{Boolean matrix with 1 row and 1 column less than the triangle to tell if the link ratio is to be considered: 1 for yes, 0 for no}
}
\value{
A list containing the following objects:
\itemize{
  \item{triangle: the input triangle }
  \item{developedTriangle: the developed triangle}
  \item{lambdas: the passing coefficients}
  \item{ibnrByAccidentYear: ibnr reserve by accident year}
  \item{ibnr: total ibnr reserve}}
}
\description{
\code{ChainLadder} applies the Chain Ladder method to a cumulated claim triangle.
}
\details{
Missing values are handled. There just need to be replaced by a NA.
}
\examples{
outputCL <- ChainLadder(triangleExampleEngland)

}
