% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChainLadderHyp.R
\name{ChainLadderHyp}
\alias{ChainLadderHyp}
\title{Verify chain ladder hypothesis}
\usage{
ChainLadderHyp(triangle, weight = NA)
}
\arguments{
\item{triangle}{Cumulated triangle as a matrix}

\item{weight}{Boolean matrix the same size of the triangle to tell if the valueis to be considered: 1 for yes, 0 for no. First column is not considered}
}
\value{
values and a plot to confirm the hypothesis
}
\description{
\code{ChainLadderHype} verifies that there is no inflation by ploting the residuals by calendar year.
}
\examples{
ChainLadderHyp(triangleExampleEngland)

}
