% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cook.R
\name{Cook}
\alias{Cook}
\title{Cook's distance}
\usage{
Cook(triangle)
}
\arguments{
\item{triangle}{Cumulated triangle as a matrix}
}
\value{
a data.frame with Cook's distances as well as a plot displaying them
}
\description{
\code{Cook} plots Cook's distance for each age to age factor relatively to the linear model they belong.
}
\examples{
Cook(triangleExampleEngland)

}
