% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mack93Variance.R
\name{Mack93Variance}
\alias{Mack93Variance}
\title{Mack93 variances}
\usage{
Mack93Variance(triangle, weight = NA)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix}

\item{weight}{Boolean matrix with 1 row and 1 column less than the triangle to tell if the link ratio is to be considered: 1 for yes, 0 for no}
}
\value{
A list containing the following objects:
\itemize{
  \item{coeffVar: variances on coefficients }
  \item{ibnrVar: variances on ibnr by accident year}
  \item{ibnrVarTot: total variance on ibnr}}
}
\description{
\code{Mack93Variance} computes the variances coefficients for each devellopment year
}
\details{
Implementation of the technique for error estimation developed by Mack in its 1993 article.
}
\examples{
error <- Mack93Variance(triangleExampleEngland)

}
