% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mack93Variance.R
\name{Mack93Variance}
\alias{Mack93Variance}
\title{Mack93 variances}
\usage{
Mack93Variance(triangle, weight = NA)
}
\arguments{
\item{triangle}{Cumulated triangle as a matrix. The matrix must have same dimension on row and columns}

\item{weight}{Boolean matrix the same size of the triangle to tell if the value is to be considered: 1 for yes, 0 for no. First column is not considered}
}
\value{
A list containing the following objects:
\itemize{
  \item{coeffVar: variances on coefficients }
  \item{ibnrVar: variances on ibnr by accident year}
  \item{ibnrVarTot: total variance on ibnr}}
}
\description{
\code{Mack93Variance} computes the variances coefficients for each development year
}
\details{
Implementation of the technique for error estimation developed by Mack in its 1993 article.
}
\examples{
error <- Mack93Variance(triangleExampleEngland)

}
