% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MackFirstHyp.R
\name{MackFirstHyp}
\alias{MackFirstHyp}
\title{Verify first Mack hypothesis}
\usage{
MackFirstHyp(triangle, mode = "separate")
}
\arguments{
\item{triangle}{Cumulated triangle as a matrix}

\item{mode}{Character. Plotting mode:
\itemize{
  \item{separate: Separate the plots by development period}
  \item{allinone: Keep all lines in one graph}}}
}
\value{
a data.frame with the data, along with a plot to confirm the hypothesis
}
\description{
\code{MackFirstHyp} verifies first Mack hypothesis about the linearity of claims.
}
\examples{
MackFirstHyp(triangleExampleEngland)

}
