% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MackSecondHyp.R
\name{MackSecondHyp}
\alias{MackSecondHyp}
\title{Verify second Mack hypothesis}
\usage{
MackSecondHyp(triangle)
}
\arguments{
\item{triangle}{Undevelopped triangle as a matrix}
}
\value{
a plot to confirm the hypothesis
}
\description{
\code{MackSecondHyp} verifies second Mack hypothesis about the independance of residuals.
}
\examples{
MackFirstHyp(triangleExampleEngland)

}
